applicationfiles/inventory/requirements.txt
aioredis==2.0.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.0.12
cleo==1.0.0a4
click==8.1.3
commonmark==0.9.1
crashtest==0.3.1
defusedxml==0.7.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
h11==0.13.0
hiredis==2.0.0
httpie==3.2.1
httptools==0.4.0
idna==3.3
multidict==6.0.2
packaging==21.3
pptree==3.1
pydantic==1.9.0
Pygments==2.12.0
pylev==1.4.0
pyparsing==3.0.9
PySocks==1.7.1
python-decouple==3.6
python-dotenv==0.20.0
python-ulid==1.1.0
PyYAML==6.0
redis==4.3.1
redis-om==0.0.27
requests==2.28.0
requests-toolbelt==0.9.1
rich==12.4.4
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
types-redis==4.2.3
types-six==1.16.15
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
cmake==3.24.1.1
